+++ CHANGELOG.md
@@ -7,11 +7,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Deprecated
+
+* `attr::charset_utf_8()` and `elt::meta_charset_utf_8()`
+  (renamed to `attr::charset_utf8()` and `elt::meta_charset_utf8()`)
+
 ### Added
 
 * `attr::none()` for conditional rendering of an attribute
 * implement `Default` for `Element` and `Attribute`
 * `attr::disabled()` and `attr::autofocus()`
+* `attr::charset_utf8()` and `elt::meta_charset_utf8()`
+
 
 ## [1.4.0] - 2024-12-13